pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "aws_doc_sdk_examples_tools"
version = "0.0.1"
dependencies = [
    'pathspec==0.11.2',
    'PyYAML==6.0.1',
    'yamale==4.0.4',
]

[tool.setuptools.packages]
find = {}